Functions improvements

+++ src/tasks/functions/generate-consumer-contract.ts
@@ -3,9 +3,19 @@ import fs from "fs";
 
 export const generateConsumerContract: ActionType<{
 }> = async (taskArgs, hre) => {
-    const src = `${hre.config.paths.root}/node_modules/@chainlink/hardhat-chainlink/contracts/FunctionsConsumer.sol`;
-    const dst = `${hre.config.paths.root}/contracts/FunctionsConsumer.sol`;
+    const dstDir = `${hre.config.paths.root}/contracts`
+    const dst = `${dstDir}/FunctionsConsumer.sol`;
+
+    if (!fs.existsSync(dstDir)) {
+        fs.mkdirSync(dstDir);
+    }
+    if (fs.existsSync(dst)) {
+        throw "Contract file exists: " + dst
+    }
+
+    const src = `${hre.config.paths.root}/node_modules/@chainlink/hardhat-chainlink/contracts/FunctionsConsumer.sol`;    
     fs.copyFileSync(src, dst);
-    console.log(`FunctionsConsumer.sol has to be saved as ${dst}`);
+    
+    console.log(`FunctionsConsumer.sol is saved as ${dst}, now compiling...`);
     await hre.run("compile");
 }
